pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "savrptw"
version = "0.1.0"
description = "Safety-Aware VRPTW for Indian quick-commerce — reference implementation for Wiley OR submission."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "SA-VRPTW project" }]

dependencies = [
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "numpy>=1.26",
  "pandas>=2.1",
  "scipy>=1.11",
  "networkx>=3.2",
  "osmnx>=1.9",
  "shapely>=2.0",
  "pyproj>=3.6",
  "scikit-learn>=1.4",
  "pulp>=2.8",
  "folium>=0.15",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "tqdm>=4.66",
]

[project.optional-dependencies]
gurobi = ["gurobipy>=11.0"]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.3",
  "black>=24.1",
  "mypy>=1.8",
]

[project.scripts]
savrptw = "savrptw.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
savrptw = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
markers = [
  "slow: tests that invoke an MILP or run > 2 s; excluded by default via -m 'not slow'",
]
